create prover crate
