+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
   "examples",
   "miden",
   "processor",
+  "prover",
   "stdlib",
   "tools",
   "verifier"
