+++ prover/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "miden-prover"
+version = "0.1.0"
+description = "Miden VM prover"
+authors = ["miden contributors"]
+readme = "README.md"
+license = "MIT"
+repository = "https://github.com/maticnetwork/miden"
+categories = ["compilers", "no-std"]
+keywords = ["prover", "proof", "miden", "zkp"]
+edition = "2021"
+rust-version = "1.60"
+
+[dependencies]
+air = { package = "miden-air", path = "../air", version = "0.2", default-features = false }
+processor = { package = "miden-processor", path = "../processor", version = "0.2", default-features = false }
+prover = { package = "winter-prover", version = "0.4", default-features = false }
+log = { version = "0.4", default-features = false }
+vm-core = { package = "miden-core", path = "../core", version = "0.2", default-features = false }
+
+[features]
+concurrent = ["prover/concurrent", "std"]
+default = ["std"]
+std = ["air/std", "processor/std", "prover/std", "log/std", "vm-core/std"]
